requirements.txt
flask
openai
python-dotenv
pymongo[srv]
flask-cors
dnspython